[Package] remove SnappySetKit lib
there is no reason to differentiate the libs, SnapKit gets pulled regardless of which lib you install. The code is small and the lib is static, so if its not used it's optimized away. A second lib only complicates setup and adds work for the linker since its linked as separate lib. If SMP ever supports pulling only deps of actually used libs, this might make sense.